+++ GraphicsFramework/src/platform/directx12/Dx12VertexBuffer.cpp
@@ -10,7 +10,7 @@ Dx12VertexBuffer::Dx12VertexBuffer(size_t size, BufferUsage usage, const void* d
 	{
 		DXCALL(D3DCreateBlob(size, m_buffer_cpu.ReleaseAndGetAddressOf()));
 		memcpy(m_buffer_cpu->GetBufferPointer(), data, size);
-		m_buffer_gpu = CreateDefaultBuffer(data, size, m_upload_buffer);
+		m_buffer_gpu = CreateDefaultBuffer(data, (unsigned int)size, m_upload_buffer);
 	}
 }
 
@@ -22,7 +22,7 @@ Dx12VertexBuffer::~Dx12VertexBuffer()
 void Dx12VertexBuffer::SetData(const void* data)
 {
 	assert(!m_buffer_gpu); // once set, data cannot be set again (for now)
-	m_buffer_gpu = CreateDefaultBuffer(data, m_size, m_upload_buffer);
+	m_buffer_gpu = CreateDefaultBuffer(data, (unsigned int)m_size, m_upload_buffer);
 }
 
 void Dx12VertexBuffer::Bind() const
